qqbot/pyproject.toml
[project]
name = "qqbot"
version = "0.1.0"
description = "qqbot"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6"
]

[tool.nonebot]
plugin_dirs = []
builtin_plugins = ["echo"]


[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
